LayoutTests/fast/dom/timer-throttling-hidden-page-2-expected.txt

Tests that DOM timers gets throttled on hidden pages once they reach the max nesting level

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS throttledCount > maxNestingLevel / 2 is true
PASS unthrottledCount > maxNestingLevel / 2 is true
PASS successfullyParsed is true

TEST COMPLETE
